utils.js: add observeAttributes to waitForChildAdd
You must now specify whether waitForChildAdd() should check the
predicate on attribute changes. None of its callers depended on this
behavior, so the callback and predicate were running hundreds of times
unnecessarily.

A new parameter, observeAttributes, has been added before abortSignal.
When it is true, you will get the old behavior. When it is false,
the predicate will not be called on attribute changes.